Moving pygame utilities
